chore: code quality improvements
improved eslint config, fixed linter issues, added husky

+++ packages/tl-runtime/src/platform/gzip.web.ts
@@ -1,9 +1,9 @@
 import { Data, Deflate, inflate } from 'pako'
 
 export function typedArrayToBuffer(arr: NodeJS.TypedArray): Buffer {
-    return ArrayBuffer.isView(arr)
-        ? Buffer.from(arr.buffer, arr.byteOffset, arr.byteLength)
-        : Buffer.from(arr)
+    return ArrayBuffer.isView(arr) ?
+        Buffer.from(arr.buffer, arr.byteOffset, arr.byteLength) :
+        Buffer.from(arr)
 }
 
 export function gzipInflate(buf: Buffer): Buffer {
@@ -18,7 +18,7 @@ class DeflateLimited extends Deflate {
     _size = 0
 
     onData(chunk: Data) {
-        this._size += chunk.length
+        this._size += (chunk as Uint8Array).length
 
         if (this._size > this.limit) {
             throw 'ERR_SIZE'
@@ -29,9 +29,9 @@ class DeflateLimited extends Deflate {
 }
 
 export function gzipDeflate(buf: Buffer, maxRatio?: number): Buffer | null {
-    const deflator = maxRatio
-        ? new DeflateLimited(Math.floor(buf.length * maxRatio))
-        : new Deflate()
+    const deflator = maxRatio ?
+        new DeflateLimited(Math.floor(buf.length * maxRatio)) :
+        new Deflate()
 
     try {
         deflator.push(buf, true)
